refactor(cannon): make spawn cannon generic by reading cannons from ModuleData/CustomXml/cannons.xml
