--- src/DellarteDellaGuerra.Infrastructure/Cannon/Mission/Siege/Spawn/Falconet.cs
@@ -1,15 +1,0 @@
-﻿using TaleWorlds.Core;
-using TaleWorlds.ObjectSystem;
-
-namespace DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn
-{
-    public class Falconet : SpawnableArtilleryRangedSiegeWeapon
-    {
-        private const string SiegeEngineId = "falconet";
-
-        public override SiegeEngineType GetSiegeEngineType()
-        {
-            return MBObjectManager.Instance.GetObject<SiegeEngineType>(SiegeEngineId);
-        }
-    }
-}
