--- src/DellarteDellaGuerra.Infrastructure/SiegeEngines/FalconetFactory.cs
@@ -1,19 +1,0 @@
-using System;
-using DellarteDellaGuerra.Infrastructure.Cannon.Mission.Siege.Spawn;
-using DellarteDellaGuerra.Infrastructure.SiegeEngines.Port;
-using TaleWorlds.MountAndBlade;
-
-namespace DellarteDellaGuerra.Infrastructure.SiegeEngines;
-
-public class FalconetFactory : ICannonFactory
-{
-    public Type CannonScriptType => typeof(Falconet);
-
-    public SpawnableArtilleryRangedSiegeWeapon CreateCannon() => new Falconet();
-
-    public void ConfigureSpawner(SpawnerEntityMissionHelper helper)
-    {
-        var cannon = helper.SpawnedEntity.GetFirstScriptInFamilyDescending<Falconet>();
-        // Configuration logic
-    }
-}
